* Fixed compiler warnings about size_t using GCC 4.2.1.

+++ include/cybergarage/upnp/ActionList.h
@@ -47,7 +47,7 @@ public:
   // Methods
   ////////////////////////////////////////////////
   
-  Action *getAction(int n) {
+  Action *getAction(size_t n) {
     return (Action*)get(n);
   }
